+++ tests/Application/Actions/Titles/ViewTitlesActionTest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tests\Application\Actions\Titles;
+
+use App\Application\Actions\Titles\ViewTitlesAction;
+use PHPUnit\Framework\TestCase;
+
+class ViewTitlesActionTest extends TestCase
+{
+
+}